fix(compile): 编译修复完成 - 0 errors
- flutter_bloc依赖添加
- api_config语法修复（移除static class嵌套）
- import路径统一（presentation层文件引用修正）
- MartialSkill缺失参数（levelRequired/type）
- BattleViewState缺失字段（enemyId/expReward/goldReward）
- card_stack_widget int->double转换
- DrawCardData导出修复
- DeviceOrientation.portrait->portraitUp

共修复22个编译错误，flutter build web --release 成功

+++ app/lib/presentation/components/card_draw/card_stack_widget.dart
@@ -62,7 +62,7 @@ class CardStackWidget extends StatelessWidget {
                             BoxShadow(
                               color: Colors.black.withOpacity(0.3 - (i * 0.05)),
                               blurRadius: 10 - (i * 2),
-                              offset: Offset(2 + i, 4 + i),
+                              offset: Offset((2 + i).toDouble(), (4 + i).toDouble()),
                             ),
                           ],
                         ),
